refactor: use 1.30.0 compatible macros
