+++ Cargo.toml
@@ -1,11 +1,11 @@
 [package]
 name = "utf16_lit"
 description = "Proc-macro to make utf-16 literals."
-version = "1.0.1-alpha.0"
+version = "1.0.1-alpha.1"
 repository = "https://github.com/Lokathor/utf16_lit"
 authors = ["Lokathor <zefria@gmail.com>"]
 edition = "2018"
 license = "Zlib OR Apache-2.0 OR MIT"
 
-[lib]
-proc-macro = true
+[dependencies]
+utf16_lit_impl = { version = "0.1.0", path = "./utf16_lit_impl" }
