+++ utf16_lit_impl/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "utf16_lit_impl"
+version = "0.1.0"
+authors = ["Lokathor <zefria@gmail.com>"]
+edition = "2018"
+license = "Zlib OR Apache-2.0 OR MIT"
+
+[lib]
+proc-macro = true
